newQuestion default useEffect and custom hook
